Better support for initialisms in spice names
Right now all uppercase characters are being treated as word boundaries, which is sub-optimal.
It causes property names like "localURL" to become "Local U R L", necessitating overrides
for all the properties.

This new implementation keeps multiple uppercase characters together as a single word, and splits
instead on the boundary where an uppercase character is introduced after a lowercase one.

This does a better job of preserving initialisms, but doesn't properly handle initialisms at the start of
the spice name. I'm not sure a great solution for that, so clients should just override those where appropriate.